Document contract: strict keys, one float widening, empty tables, carriers, JSON specials, parity corpus (#214)
* Document contract: strict keys, one float widening, empty tables, carriers, JSON specials, parity corpus

Closes #198: duplicate keys are refused at every level. Tomlyn's TomlTable
binding ignores DuplicateKeyHandling and keeps the last value, so the reader
validates the syntax tree first (Tomlyn's semantic pass catches root, nested,
inline, dotted and header duplicates); the typed manifest context sets the
option as well. The manifest refuses unknown keys through [TomlExtensionData]
at the root, in [build] and on each profile, and refuses document_format =
"blob" and pack = true until a writer exists.

Closes #199: an empty generic table is written `key = {}` and an empty array
of tables `key = []`, at value position, which is the only form the reader can
restore them from; read → write is identity again and the spec text says so.

Closes #200: one float32 widening (CanonicalFloat.Widen, shortest decimal)
for tables, inline tables and the transform codec; 0.1f is 0.1 everywhere.

Closes #210: an override carrier must name its Parent, the Parent must exist,
and a carrier may not declare a Guid; a one-half asset reference cannot be
written.

Closes #211: CanonicalJson is the one canonical-model → JSON converter, used
by the config compile and the prefab bake (the bake keeps its built-path
substitution as a hook). inf and nan are refused with the source and key
named; the integral-float collapse is documented and pinned. ConfigDocument
gains tests.

Towards #209: a parity corpus under Paradise.Assets.Documents.Test/Fixtures/
parity, written by the C# writer (floats at every boundary, strings and key
quoting, structure, a prefab with carriers and null slots), pinned as a fixed
point of read → write. The Python consumer is the addon repo's half.


* inf/nan refusal lands on the importer error list, not as an exception

The importer contract is an error on the list, never an exception out of
Import; CanonicalJson's FormatException escaped both the prefab bake and the
config importer, so a json-profile build of legal TOML still died. The bake
now reports per component, naming the object and the component; the config
importer reports the source and key path. Tests at both levels.

Review nits: the corpus-presence guard is a floor, not an exact count; the
verifier's walk comment names CanonicalJson; RejectUnknown lists every
unknown key at once.


---------

+++ src/Paradise.Assets.Documents/LocalTransformCodec.cs
@@ -20,22 +20,22 @@ public static class LocalTransformCodec
             Vector(data.Value(WellKnownComponents.Scale), Vector3.One));
     }
 
-    /// <summary>All three fields, in canonical order; widens float32 bit-exactly, unlike <c>CanonicalTomlTable.Add</c> (issue #200).</summary>
+    /// <summary>All three fields, in canonical order. The floats are handed to the table as float32 so its widening (<see cref="CanonicalFloat"/>) spells them.</summary>
     public static PrefabComponent Write(LocalTransform transform)
         => new(WellKnownComponents.TransformId, WellKnownComponents.TransformType,
             new CanonicalTomlTable
             {
                 {
                     WellKnownComponents.Position,
-                    new object[] { (double)transform.Position.X, (double)transform.Position.Y, (double)transform.Position.Z }
+                    new object[] { transform.Position.X, transform.Position.Y, transform.Position.Z }
                 },
                 {
                     WellKnownComponents.Rotation,
-                    new object[] { (double)transform.Rotation.X, (double)transform.Rotation.Y, (double)transform.Rotation.Z, (double)transform.Rotation.W }
+                    new object[] { transform.Rotation.X, transform.Rotation.Y, transform.Rotation.Z, transform.Rotation.W }
                 },
                 {
                     WellKnownComponents.Scale,
-                    new object[] { (double)transform.Scale.X, (double)transform.Scale.Y, (double)transform.Scale.Z }
+                    new object[] { transform.Scale.X, transform.Scale.Y, transform.Scale.Z }
                 },
             });
